requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
httpx
pytest
python-dotenv
google-generativeai
google-auth